+++ lib/src/sc_builder.dart
@@ -139,11 +139,11 @@ class ScBuilder {
   }
 
   void _applyBuildersToChildScopes(Scope scope) {
-    if (!bluePrint.shouldDigInto(scope)) {
+    _applyChildBuilders(scope);
+
+    if (!bluePrint.shouldProcessChildrenOf(scope)) {
       return;
     }
-
-    _applyChildBuilders(scope);
 
     for (final childScope in scope.children) {
       _applyBuildersToChildScopes(childScope);