+++ test/scm_test.dart
@@ -956,7 +956,7 @@ void main() {
           produce: (components, previousProduct) {
             ScBuilderBluePrint(
               key: 'builder',
-              shouldDigInto: (scope) => const [
+              shouldProcessChildrenOf: (scope) => const [
                 'example',
               ].contains(scope.key),
               addNodes: ({required hostScope}) {